vids/pyproject.toml
[tool.poetry]
name = "vids"
version = "0.1.0"
description = ""
authors = ["will-s-stone <will@williamstone.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
ffmpeg-python = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
